+++ core/src/test-core/java/com/radixdlt/rev1/MockedMempoolStateComputerModule.java
@@ -76,12 +76,12 @@ import com.radixdlt.consensus.bft.VertexStoreState;
 import com.radixdlt.environment.EventDispatcher;
 import com.radixdlt.ledger.CommittedTransactionsWithProof;
 import com.radixdlt.ledger.LedgerUpdate;
+import com.radixdlt.ledger.MockExecuted;
 import com.radixdlt.ledger.StateComputerLedger;
 import com.radixdlt.mempool.Mempool;
 import com.radixdlt.mempool.MempoolAdd;
 import com.radixdlt.mempool.MempoolMaxSize;
 import com.radixdlt.mempool.MempoolRejectedException;
-import com.radixdlt.modules.MockExecuted;
 import com.radixdlt.monitoring.SystemCounters;
 import com.radixdlt.targeted.mempool.SimpleMempool;
 import com.radixdlt.transactions.Transaction;
