+++ core/src/test-core/java/com/radixdlt/rev1/MockedStateComputer.java
@@ -82,10 +82,10 @@ import com.radixdlt.crypto.Hasher;
 import com.radixdlt.environment.EventDispatcher;
 import com.radixdlt.ledger.CommittedTransactionsWithProof;
 import com.radixdlt.ledger.LedgerUpdate;
+import com.radixdlt.ledger.MockExecuted;
 import com.radixdlt.ledger.StateComputerLedger;
 import com.radixdlt.ledger.StateComputerLedger.StateComputer;
 import com.radixdlt.mempool.MempoolAdd;
-import com.radixdlt.modules.MockExecuted;
 import com.radixdlt.transactions.Transaction;
 import java.util.List;
 import java.util.Map;
