+++ core/src/test-core/java/com/radixdlt/rev1/MockedStateComputerWithEpochs.java
@@ -73,11 +73,11 @@ import com.radixdlt.crypto.Hasher;
 import com.radixdlt.environment.EventDispatcher;
 import com.radixdlt.ledger.CommittedTransactionsWithProof;
 import com.radixdlt.ledger.LedgerUpdate;
+import com.radixdlt.ledger.MockExecuted;
 import com.radixdlt.ledger.StateComputerLedger.ExecutedTransaction;
 import com.radixdlt.ledger.StateComputerLedger.StateComputer;
 import com.radixdlt.ledger.StateComputerLedger.StateComputerResult;
 import com.radixdlt.mempool.MempoolAdd;
-import com.radixdlt.modules.MockExecuted;
 import com.radixdlt.transactions.Transaction;
 import java.util.List;
 import java.util.Objects;
